rewrote manual mappers to MapStruct. Fixed price type from Double to BigDecimal

+++ src/main/java/com/best_store/right_bite/dto/catalog/ProductFilterRequest.java
@@ -2,6 +2,8 @@ package com.best_store.right_bite.dto.catalog;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProductFilterRequest {
 
@@ -9,8 +11,8 @@ public class ProductFilterRequest {
 
     private String brand;
 
-    private Double priceFrom;
-    private Double priceTo;
+    private BigDecimal priceFrom;
+    private BigDecimal priceTo;
 
     private Character aZ;
